LPS-64035 Fix test
