+++ modules/apps/forms-and-workflow/dynamic-data-lists/dynamic-data-lists-test/src/testIntegration/java/com/liferay/dynamic/data/lists/helper/DDLRecordSetTestHelper.java
@@ -44,6 +44,13 @@ public class DDLRecordSetTestHelper {
 	public DDLRecordSet addRecordSet(DDMStructure ddmStructure)
 		throws Exception {
 
+		return addRecordSet(
+			ddmStructure, DDLRecordSetConstants.SCOPE_DYNAMIC_DATA_LISTS);
+	}
+
+	public DDLRecordSet addRecordSet(DDMStructure ddmStructure, int scope)
+		throws Exception {
+
 		Map<Locale, String> nameMap = new HashMap<>();
 
 		nameMap.put(LocaleUtil.US, RandomTestUtil.randomString());
@@ -54,8 +61,8 @@ public class DDLRecordSetTestHelper {
 		return DDLRecordSetLocalServiceUtil.addRecordSet(
 			TestPropsValues.getUserId(), _group.getGroupId(),
 			ddmStructure.getStructureId(), null, nameMap, null,
-			DDLRecordSetConstants.MIN_DISPLAY_ROWS_DEFAULT,
-			DDLRecordSetConstants.SCOPE_DYNAMIC_DATA_LISTS, serviceContext);
+			DDLRecordSetConstants.MIN_DISPLAY_ROWS_DEFAULT, scope,
+			serviceContext);
 	}
 
 	public DDLRecordSet updateRecordSet(
